requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
requests==2.32.3
flask-sqlalchemy==3.1.1
pymysql==1.1.1
cryptography
openai==0.28
flask-migrate